+++ cmd/frontend/internal/app/one_click_export.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/sourcegraph/log"
 
-	oce "github.com/sourcegraph/sourcegraph/cmd/frontend/oneclickexport"
+	oce "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/oneclickexport"
 	"github.com/sourcegraph/sourcegraph/internal/auth"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 )
