+++ cmd/frontend/internal/oneclickexport/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
         "export.go",
         "processor.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/oneclickexport",
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/oneclickexport",
     tags = [TAG_SEARCHSUITE],
     visibility = ["//visibility:public"],
     deps = [
